fix(User-Management): Move common code to base class
