+++ dj_users/tests/test_user_login.py
@@ -2,31 +2,16 @@
 import pytest
 
 from django.urls import reverse
-from django.contrib.auth import get_user_model
 
+from .base_classes import UserLoginLogoutBase
 
 pytestmark = [pytest.mark.django_db]
 
 
-# noinspection HardcodedPassword
-class TestUserLogin:
+class TestUserLogin(UserLoginLogoutBase):
     """Test login of users."""
 
-    USER_MODEL = get_user_model()
-
     LOGIN_URL = reverse("account_login")
-
-    USERNAME = "testuser"
-    USER_EMAIL = "test_user@provided.com"
-    PASSWORD = "testpassword123"
-
-    @pytest.fixture(autouse=True)
-    def setup_method(self, db) -> None:
-        """Set up a user for testing login."""
-        # Set up a user for testing login
-        self.USER_MODEL.objects.create_user(
-            self.USERNAME, self.USER_EMAIL, self.PASSWORD
-        )
 
     def test_successful_login(self, client) -> None:
         """Test a successful login."""
